requirements.txt
numpy<2
scipy
opencv-python<4.12
Pillow
tqdm
matplotlib
seaborn
torch
torchvision
timm
lpips
pytubefix
scikit-image
python-dotenv
pydantic
transformers
qwen-vl-utils
accelerate
sentencepiece
protobuf
ultralytics
anthropic
requests
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/sam3.git
git+https://github.com/DepthAnything/Depth-Anything-V2.git
einops
decord
pycocotools
triton-windows<3.6; platform_system == "Windows"
openai
pandas
openpyxl# VQA2 evaluation metrics
rouge-score
nltk
bert-score